+++ bin/lint.sh
@@ -5,4 +5,4 @@ set -e
 bin_folder="`dirname "$0"`"
 root_folder="$bin_folder/.."
 
-flux-js-lint "$root_folder"
+js-lint "$root_folder"