+++ bin/publish.sh
@@ -6,9 +6,9 @@ bin_folder="`dirname "$0"`"
 root_folder="$bin_folder/.."
 
 name="`basename "$(realpath "$root_folder")"`"
-host="${FLUX_PUBLISH_DOCKER_HOST:=}"
+host="${PUBLISH_DOCKER_HOST:=}"
 host_with_slash="${host}${host:+/}"
-image="$host_with_slash$FLUX_PUBLISH_DOCKER_USER/$name"
+image="$host_with_slash$PUBLISH_DOCKER_USER/$name"
 tag="`$bin_folder/HOST_PATH/get-release-tag "$root_folder"`"
 
 "$bin_folder/build.sh"
@@ -16,7 +16,7 @@ tag="`$bin_folder/HOST_PATH/get-release-tag "$root_folder"`"
 "$bin_folder/HOST_PATH/tag-release" "$root_folder"
 "$bin_folder/HOST_PATH/create-github-release" "$root_folder"
 
-export DOCKER_CONFIG="$FLUX_PUBLISH_DOCKER_CONFIG_FOLDER"
+export DOCKER_CONFIG="$PUBLISH_DOCKER_CONFIG_FOLDER"
 docker push "$image:$tag"
 docker push "$image:latest"
 unset DOCKER_CONFIG