+++ bin/upload-asset-to-github-release.mjs
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
-import { FluxShutdownHandler } from "flux-shutdown-handler/src/FluxShutdownHandler.mjs";
+import { ShutdownHandler } from "shutdown-handler/src/ShutdownHandler.mjs";
 
-const flux_shutdown_handler = await FluxShutdownHandler.new();
+const shutdown_handler = await ShutdownHandler.new();
 
 try {
     const path = process.argv[2] ?? null;
@@ -23,7 +23,7 @@ try {
 } catch (error) {
     console.error(error);
 
-    await flux_shutdown_handler.shutdown(
+    await shutdown_handler.shutdown(
         1
     );
 }