+++ src/Config/CONFIG.mjs
@@ -1,1 +1,1 @@
-export const CONFIG_ENV_PREFIX = "FLUX_PUBLISH_UTILS_";
+export const CONFIG_ENV_PREFIX = "PUBLISH_UTILS_";