+++ src/GetGithubAuthorization.mjs
@@ -1,7 +1,7 @@
+import { Config } from "config/src/Config.mjs";
 import { CONFIG_ENV_PREFIX } from "./Config/CONFIG.mjs";
-import { CONFIG_TYPE_STRING } from "flux-config/src/CONFIG_TYPE.mjs";
-import { FluxConfig } from "flux-config/src/FluxConfig.mjs";
-import { getValueProviders } from "flux-config/src/getValueProviders.mjs";
+import { CONFIG_TYPE_STRING } from "config/src/CONFIG_TYPE.mjs";
+import { getValueProviders } from "config/src/getValueProviders.mjs";
 import { GITHUB_CONFIG_TOKEN_KEY } from "./Github/GITHUB_CONFIG.mjs";
 
 export class GetGithubAuthorization {
@@ -23,7 +23,7 @@ export class GetGithubAuthorization {
      * @returns {Promise<string>}
      */
     async getGithubAuthorization() {
-        return `Basic ${btoa(await (await FluxConfig.new(
+        return `Basic ${btoa(await (await Config.new(
             await getValueProviders(
                 null,
                 CONFIG_ENV_PREFIX